requirements.txt
playwright
pillow
pytesseract
python-dotenv
opencv-python